+++ lib/ios_storekit2.dart
@@ -8,12 +8,18 @@ class IosStorekit2 {
     return IosStorekit2Platform.instance.getProducts(identifiers);
   }
 
-  Future<SK2PurchaseResult> purchase(String productId) {
-    return IosStorekit2Platform.instance.purchase(productId);
+  Future<SK2PurchaseResult> purchase(String productId,
+      {String? appAccountToken}) {
+    return IosStorekit2Platform.instance
+        .purchase(productId, appAccountToken: appAccountToken);
   }
 
   Future<List<SK2Entitlement>> getEntitlements() {
     return IosStorekit2Platform.instance.getEntitlements();
+  }
+
+  Future<SK2Storefront?> getStorefront() {
+    return IosStorekit2Platform.instance.getStorefront();
   }
 
   Future<void> restorePurchases() {